requirements.txt
flask
python-dotenv
requests
msal
flask-session